[tests] Split tests of HiFi-GAN
